# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Backend Template","version":"0.1.0"},"paths":{"/api/v1/auth/register":{"post":{"tags":["Auth"],"summary":"Register User","operationId":"register_user_api_v1_auth_register_post","parameters":[{"name":"X-Client","in":"header","required":false,"schema":{"enum":["web","mobile"],"type":"string","default":"web","title":"X-Client"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}}},"responses":{"201":{"description":"refresh_token field varies depending on the source of request. For web it is always null and being set as httponly cookie, however any other platform gets access and refresh tokens in response body. It is made to protect web from xss and csrf attacks","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenPair"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"429":{"description":"Too Many Requests"},"409":{"description":"User with provided credentials already exists"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Authenticate user and issue tokens","operationId":"login_user_api_v1_auth_login_post","parameters":[{"name":"X-Client","in":"header","required":false,"schema":{"enum":["web","mobile"],"type":"string","default":"web","title":"X-Client"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}}},"responses":{"200":{"description":"refresh_token field varies depending on the source of request. For web it is always null and being set as httponly cookie, however any other platform gets access and refresh tokens in response body. It is made to protect web from xss and csrf attacks","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenPair"}}}},"401":{"description":"Wrong credentials"},"403":{"description":"Forbidden"},"429":{"description":"Too Many Requests"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["Auth"],"summary":"Logout","operationId":"logout_api_v1_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Logout Api V1 Auth Logout Post"}}}},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"429":{"description":"Too Many Requests"}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/refresh":{"post":{"tags":["Auth"],"summary":"Rotate tokens","operationId":"refresh_tokens_api_v1_auth_refresh_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"X-CSRF-Token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Must only be passed for requests from browsers","title":"X-Csrf-Token"},"description":"Must only be passed for requests from browsers"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenPair"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"429":{"description":"Too Many Requests"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/me":{"get":{"tags":["Users"],"summary":"Get user account info","operationId":"profile_api_v1_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}},"401":{"description":"Not authorized"}},"security":[{"HTTPBearer":[]}]},"patch":{"tags":["Users"],"summary":"Update user info","operationId":"update_profile_api_v1_users_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPatch"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}},"401":{"description":"Not authorized"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/users/me/avatar/presign":{"post":{"tags":["Users"],"summary":"Create presigned URL for avatar upload","operationId":"create_avatar_presigned_upload_api_v1_users_me_avatar_presign_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvatarPresignRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvatarPresignResponse"}}}},"401":{"description":"Not authorized"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/users/me/avatar/confirm":{"post":{"tags":["Users"],"summary":"Confirm uploaded avatar and attach it to profile","operationId":"confirm_avatar_upload_api_v1_users_me_avatar_confirm_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvatarConfirmRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}},"401":{"description":"Not authorized"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/users/me/avatar":{"delete":{"tags":["Users"],"summary":"Remove current user avatar","operationId":"delete_avatar_api_v1_users_me_avatar_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}},"401":{"description":"Not authorized"}},"security":[{"HTTPBearer":[]}]}},"/api/v1/languages":{"get":{"tags":["Languages"],"summary":"List languages with search","operationId":"list_languages_api_v1_languages_get","parameters":[{"name":"query","in":"query","required":false,"schema":{"type":"string","maxLength":50,"default":"","title":"Query"}},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":50,"minimum":1},{"type":"null"}],"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LanguageModel"},"title":"Response List Languages Api V1 Languages Get"}}}},"401":{"description":"Not authorized"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admins/users/":{"get":{"tags":["Admins"],"summary":"List users with filters and search (cursor pagination)","operationId":"list_users_api_v1_admins_users__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"banned","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"description":"Filter by banned status","title":"Banned"},"description":"Filter by banned status"},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search by username or email","title":"Search"},"description":"Search by username or email"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Page size","default":50,"title":"Limit"},"description":"Page size"},{"name":"cursor","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Opaque cursor","title":"Cursor"},"description":"Opaque cursor"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CursorPage_UserModel_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admins/users/{user_id}/ban":{"post":{"tags":["Admins"],"summary":"Ban or unban a user","operationId":"set_ban_api_v1_admins_users__user_id__ban_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}},{"name":"banned","in":"query","required":false,"schema":{"type":"boolean","description":"Set to false to remove ban","default":true,"title":"Banned"},"description":"Set to false to remove ban"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admins/users/{user_id}/roles":{"put":{"tags":["Admins"],"summary":"Assign roles to user","operationId":"set_roles_api_v1_admins_users__user_id__roles_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRolesUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admins/stats/summary":{"get":{"tags":["Admins"],"summary":"Get users summary counters","operationId":"users_summary_api_v1_admins_stats_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserStatsSummary"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/admins/stats/registrations":{"get":{"tags":["Admins"],"summary":"Get graph data for new registrations by days","operationId":"registrations_api_v1_admins_stats_registrations_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","description":"Number of days back to retrieve data for","default":30,"title":"Days"},"description":"Number of days back to retrieve data for"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RegistrationsGraph"},"title":"Response Registrations Api V1 Admins Stats Registrations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/health":{"get":{"summary":"Liveness","operationId":"liveness_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/ping":{"get":{"summary":"Liveness","operationId":"liveness_api_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/ready":{"get":{"summary":"Readiness","operationId":"readiness_api_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AvatarConfirmRequest":{"properties":{"object_key":{"type":"string","maxLength":1024,"minLength":1,"title":"Object Key"}},"type":"object","required":["object_key"],"title":"AvatarConfirmRequest"},"AvatarPresignRequest":{"properties":{"filename":{"type":"string","maxLength":255,"minLength":1,"title":"Filename"},"content_type":{"type":"string","enum":["image/jpeg","image/png","image/webp"],"title":"Content Type"}},"type":"object","required":["filename","content_type"],"title":"AvatarPresignRequest"},"AvatarPresignResponse":{"properties":{"object_key":{"type":"string","minLength":1,"title":"Object Key"},"upload_url":{"type":"string","minLength":1,"title":"Upload Url"},"public_url":{"type":"string","minLength":1,"title":"Public Url"},"expires_in":{"type":"integer","exclusiveMinimum":0.0,"title":"Expires In"}},"type":"object","required":["object_key","upload_url","public_url","expires_in"],"title":"AvatarPresignResponse"},"CursorPage_UserModel_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/UserModel"},"type":"array","title":"Items"},"next_cursor":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next Cursor"}},"type":"object","required":["items"],"title":"CursorPage[UserModel]"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LanguageModel":{"properties":{"code":{"type":"string","maxLength":2,"minLength":2,"title":"Code"},"name_ru":{"type":"string","title":"Name Ru"},"name_en":{"type":"string","title":"Name En"}},"type":"object","required":["code","name_ru","name_en"],"title":"LanguageModel"},"RegistrationsGraph":{"properties":{"day":{"type":"string","format":"date","title":"Day"},"count":{"type":"integer","title":"Count"}},"type":"object","required":["day","count"],"title":"RegistrationsGraph"},"TokenPair":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Refresh Token"}},"type":"object","required":["access_token"],"title":"TokenPair"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"User e-mail"},"password":{"type":"string","title":"Password","description":"User password"}},"type":"object","required":["email","password"],"title":"UserLogin","description":"Credentials used for user login."},"UserModel":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"email":{"type":"string","format":"email","title":"Email","description":"User e-mail"},"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username","description":"User's display name"},"avatar_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar Key","description":"Storage key of the avatar object"},"avatar_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar Url","description":"Public URL of the avatar object"},"is_onboarded":{"type":"boolean","title":"Is Onboarded"},"banned":{"type":"boolean","title":"Banned"},"role_slugs":{"items":{"type":"string"},"type":"array","title":"Role Slugs","description":"User's roles."},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["created_at","updated_at","id","email","is_onboarded","banned"],"title":"UserModel","description":"User account representation."},"UserPatch":{"properties":{"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username","description":"User's display name"}},"type":"object","title":"UserPatch"},"UserRegister":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"User e-mail"},"password":{"type":"string","title":"Password","description":"User password"},"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username","description":"User's display name"}},"type":"object","required":["email","password"],"title":"UserRegister","description":"Data required for user registration."},"UserRolesUpdate":{"properties":{"roles":{"items":{"type":"string"},"type":"array","title":"Roles","description":"Role slugs to assign"}},"type":"object","title":"UserRolesUpdate"},"UserStatsSummary":{"properties":{"generated_at":{"type":"string","format":"date-time","title":"Generated At"},"total_users":{"type":"integer","minimum":0.0,"title":"Total Users"},"onboarded_users":{"type":"integer","minimum":0.0,"title":"Onboarded Users"},"banned_users":{"type":"integer","minimum":0.0,"title":"Banned Users"},"registered_last_24h":{"type":"integer","minimum":0.0,"title":"Registered Last 24H"}},"type":"object","required":["generated_at","total_users","onboarded_users","banned_users","registered_last_24h"],"title":"UserStatsSummary"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","description":"Access token must be passed as Bearer to authorize request","scheme":"bearer"}}}}